Move `misc::Mention{able}` into their own `model/mention` module
